continue moving server pieces into their final positions
prettify server files

fix ent import path in worker.js
