+++ server/.dockerignore
@@ -1,0 +1,3 @@
+*.log
+.DS_Store
+node_modules/
