+++ server/ent/hyb-ent
@@ -1,8 +1,8 @@
 #!/usr/bin/env node
 'use strict'
 
-const getData = require('./lib/get-data')
-const { strictSearch, formatData } = require('../server/lib/ent')
+const getData = require('../lib/get-data')
+const { strictSearch, formatData } = require('./index')
 
 function main() {
 	let file = process.argv[2]
