+++ server/ent/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+const ent = require('./ent')
+
+module.exports = ent
