+++ server/formats/fasta/index.js
@@ -1,0 +1,6 @@
+'use strict'
+
+const build = require('./build')
+const parse = require('./parse')
+
+module.exports = { build, parse }
