+++ server/formats/genbank-to-fasta.js
