+++ server/formats/hyb-fasta-to-nexus
@@ -1,14 +1,14 @@
 #!/usr/bin/env node
 'use strict'
 
-const getData = require('@hybsearch/get-data')
-const fastaToNexus = require('../fasta-to-nexus')
+const getData = require('../lib/get-data')
+const fastaToNexus = require('./fasta-to-nexus')
 
 function main() {
 	let file = process.argv[2]
 
 	if (!file && process.stdin.isTTY) {
-		console.error('usage: node fasta-to-nexus.js (<input> | -)')
+		console.error('usage: hyb-fasta-to-nexus (<input> | -)')
 		process.exit(1)
 	}
