+++ server/formats/hyb-genbank-to-fasta
@@ -1,20 +1,20 @@
 #!/usr/bin/env node
 'use strict'
 
-const getData = require('@hybsearch/get-data')
-const convert = require('../nexus-to-fasta')
+const getData = require('../get-data')
+const genbankToFasta = require('./genbank-to-fasta')
 
 function main() {
 	let file = process.argv[2]
 
 	if (!file && process.stdin.isTTY) {
-		console.error('usage: node convert-nexus-to-fasta.js (<input> | -)')
+		console.error('usage: hyb-genbank-to-fasta (<input> | -)')
 		process.exit(1)
 	}
 
 	return getData(file)
-		.then(convert)
-		.then(console.log.bind(console))
+		.then(genbankToFasta)
+		.then(data => console.log(data))
 		.catch(console.error.bind(console))
 }
