+++ server/formats/hyb-nexus-to-fasta
@@ -1,20 +1,20 @@
 #!/usr/bin/env node
 'use strict'
 
-const getData = require('./lib/get-data')
-const consensusTreeNewick = require('../bin/consensus-newick')
+const getData = require('../get-data')
+const convert = require('./nexus-to-fasta')
 
 function main() {
 	let file = process.argv[2]
 
 	if (!file && process.stdin.isTTY) {
-		console.error('usage: node consensus-newick.js (<input> | -)')
+		console.error('usage: hyb-nexus-to-fasta (<input> | -)')
 		process.exit(1)
 	}
 
 	return getData(file)
-		.then(consensusTreeNewick)
-		.then(data => console.log(data))
+		.then(convert)
+		.then(console.log.bind(console))
 		.catch(console.error.bind(console))
 }
