+++ server/hamdis/hyb-hamdis
@@ -1,12 +1,18 @@
 #!/usr/bin/env node
 'use strict'
 
-const getData = require('./lib/get-data')
-const hamdis = require('../server/lib/hamdis')
-const estimateGenerations = require('../server/lib/estimate-generations')
+const getData = require('../lib/get-data')
+const { hamdis, estimateGenerations } = require('./index')
 
 function main() {
-	return getData(process.argv[2])
+	let file = process.argv[2]
+
+	if (!file && process.stdin.isTTY) {
+		console.error('usage: hyb-hamdis (<input> | -)')
+		process.exit(1)
+	}
+
+	return getData(file)
 		.then(data => {
 			let stats = hamdis(data)
 			console.log('distances', stats.data)
