+++ server/hamdis/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const estimateGenerations = require('./estimate-generations')
+const hammingDistance = require('./hamming-distance')
+const hamdis = require('./hamdis')
+
+module.exports = { hamdis, hammingDistance, estimateGenerations }
