+++ server/newick/hyb-newick-to-json
@@ -1,20 +1,20 @@
 #!/usr/bin/env node
 'use strict'
 
-const getData = require('./lib/get-data')
-const mrbayes = require('../bin/mrbayes')
+const getData = require('../lib/get-data')
+const newick = require('./parser')
 
 function main() {
-	let argv = process.argv.slice(2)
-	let file = argv[0]
+	let file = process.argv[2]
 
 	if (!file && process.stdin.isTTY) {
-		console.error('usage: node mrbayes.js (<input> | -) [--quiet]')
+		console.error('usage: hyb-newick-to-json (<input> | -)')
 		process.exit(1)
 	}
 
 	return getData(file)
-		.then(data => mrbayes(data, argv))
+		.then(newick)
+		.then(d => JSON.stringify(d))
 		.then(data => console.log(data))
 		.catch(console.error.bind(console))
 }
