+++ server/newick/index.js
@@ -1,0 +1,6 @@
+'use strict'
+
+const consensusTreeToNewick = require('./consensus-to-newick')
+const parser = require('./parser')
+
+module.exports = { parse: parser, consensusTreeToNewick }
