+++ server/pipeline/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+const pipeline = require('./worker')
+
+module.exports = pipeline
