+++ server/wrappers/mrbayes/hyb-mrbayes
@@ -1,19 +1,20 @@
 #!/usr/bin/env node
 'use strict'
 
-const getData = require('./lib/get-data')
-const seqmagick = require('../bin/fasta-to-nexus')
+const getData = require('../../lib/get-data')
+const mrbayes = require('./index')
 
 function main() {
-	let file = process.argv[2]
+	let argv = process.argv.slice(2)
+	let file = argv[0]
 
 	if (!file && process.stdin.isTTY) {
-		console.error('usage: node fasta-to-nexus.js (<input> | -)')
+		console.error('usage: hyb-mrbayes (<input> | -) [--quiet]')
 		process.exit(1)
 	}
 
 	return getData(file)
-		.then(seqmagick)
+		.then(data => mrbayes(data, argv))
 		.then(data => console.log(data))
 		.catch(console.error.bind(console))
 }
