+++ server/wrappers/seq-gen/hyb-seq-gen
@@ -10,10 +10,10 @@ optional arguments:
   --generations 2
 `
 
-const getData = require('./lib/get-data')
+const getData = require('../../lib/get-data')
 const minimist = require('minimist')
 
-const seqgen = require('../bin/seq-gen')
+const seqgen = require('./index')
 
 function main() {
 	let argv = minimist(process.argv.slice(2))
